Simplify related query to remove nesting
